Moved cloud drives detection to FTP
