+++ Files.Launcher/Files.Launcher.csproj
@@ -118,6 +118,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DeviceWatcher.cs" />
+    <Compile Include="Helpers\CloudDrivesDetector.cs" />
     <Compile Include="Helpers\DisposableDictionary.cs" />
     <Compile Include="Helpers\ShellFolderHelpers.cs" />
     <Compile Include="Helpers\ShellNewMenuHelper.cs" />
