+++ Files.Launcher/MessageHandlers/NetworkDrivesHandler.cs
@@ -1,5 +1,7 @@
 ﻿using Files.Common;
+using FilesFullTrust.Helpers;
 using Microsoft.Win32;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -26,95 +28,12 @@ namespace FilesFullTrust.MessageHandlers
                     await ParseNetworkDriveOperationAsync(connection, message);
                     break;
 
-                case "GetOneDriveAccounts":
-                    try
+                case "DetectCloudDrives":
+                    var cloudDrives = await CloudDrivesDetector.DetectCloudDrives();
+                    await Win32API.SendMessageAsync(connection, new ValueSet()
                     {
-                        using var oneDriveAccountsKey = Registry.CurrentUser.OpenSubKey(@"SOFTWARE\Microsoft\OneDrive\Accounts");
-
-                        if (oneDriveAccountsKey == null)
-                        {
-                            await Win32API.SendMessageAsync(connection, new ValueSet() { { "Count", 0 } }, message.Get("RequestID", (string)null));
-                            return;
-                        }
-
-                        var oneDriveAccounts = new ValueSet();
-                        foreach (var account in oneDriveAccountsKey.GetSubKeyNames())
-                        {
-                            var accountKeyName = @$"{oneDriveAccountsKey.Name}\{account}";
-                            var displayName = (string)Registry.GetValue(accountKeyName, "DisplayName", null);
-                            var userFolder = (string)Registry.GetValue(accountKeyName, "UserFolder", null);
-                            var accountName = string.IsNullOrWhiteSpace(displayName) ? "OneDrive" : $"OneDrive - {displayName}";
-                            if (!string.IsNullOrWhiteSpace(userFolder) && !oneDriveAccounts.ContainsKey(accountName))
-                            {
-                                oneDriveAccounts.Add(accountName, userFolder);
-                            }
-                        }
-                        oneDriveAccounts.Add("Count", oneDriveAccounts.Count);
-                        await Win32API.SendMessageAsync(connection, oneDriveAccounts, message.Get("RequestID", (string)null));
-                    }
-                    catch
-                    {
-                        await Win32API.SendMessageAsync(connection, new ValueSet() { { "Count", 0 } }, message.Get("RequestID", (string)null));
-                    }
-                    break;
-
-                case "GetSharePointSyncLocationsFromOneDrive":
-                    try
-                    {
-                        using var oneDriveAccountsKey = Registry.CurrentUser.OpenSubKey(@"SOFTWARE\Microsoft\OneDrive\Accounts");
-
-                        if (oneDriveAccountsKey == null)
-                        {
-                            await Win32API.SendMessageAsync(connection, new ValueSet() { { "Count", 0 } }, message.Get("RequestID", (string)null));
-                            return;
-                        }
-
-                        var sharepointAccounts = new ValueSet();
-
-                        foreach (var account in oneDriveAccountsKey.GetSubKeyNames())
-                        {
-                            var accountKeyName = @$"{oneDriveAccountsKey.Name}\{account}";
-                            var displayName = (string)Registry.GetValue(accountKeyName, "DisplayName", null);
-                            var userFolderToExcludeFromResults = (string)Registry.GetValue(accountKeyName, "UserFolder", null);
-                            var accountName = string.IsNullOrWhiteSpace(displayName) ? "SharePoint" : $"SharePoint - {displayName}";
-
-                            var sharePointSyncFolders = new List<string>();
-                            var mountPointKeyName = @$"SOFTWARE\Microsoft\OneDrive\Accounts\{account}\ScopeIdToMountPointPathCache";
-                            using (var mountPointsKey = Registry.CurrentUser.OpenSubKey(mountPointKeyName))
-                            {
-                                if (mountPointsKey == null)
-                                {
-                                    continue;
-                                }
-
-                                var valueNames = mountPointsKey.GetValueNames();
-                                foreach (var valueName in valueNames)
-                                {
-                                    var value = (string)Registry.GetValue(@$"HKEY_CURRENT_USER\{mountPointKeyName}", valueName, null);
-                                    if (!string.Equals(value, userFolderToExcludeFromResults, StringComparison.OrdinalIgnoreCase))
-                                    {
-                                        sharePointSyncFolders.Add(value);
-                                    }
-                                }
-                            }
-
-                            foreach (var sharePointSyncFolder in sharePointSyncFolders.OrderBy(o => o))
-                            {
-                                var parentFolder = Directory.GetParent(sharePointSyncFolder)?.FullName ?? string.Empty;
-                                if (!sharepointAccounts.Any(acc => string.Equals(acc.Key, accountName, StringComparison.OrdinalIgnoreCase)) && !string.IsNullOrWhiteSpace(parentFolder))
-                                {
-                                    sharepointAccounts.Add(accountName, parentFolder);
-                                }
-                            }
-                        }
-
-                        sharepointAccounts.Add("Count", sharepointAccounts.Count);
-                        await Win32API.SendMessageAsync(connection, sharepointAccounts, message.Get("RequestID", (string)null));
-                    }
-                    catch
-                    {
-                        await Win32API.SendMessageAsync(connection, new ValueSet() { { "Count", 0 } }, message.Get("RequestID", (string)null));
-                    }
+                        { "Drives", JsonConvert.SerializeObject(cloudDrives) }
+                    }, message.Get("RequestID", (string)null));
                     break;
             }
         }
