--- Files/Enums/CloudProviders.cs
@@ -1,14 +1,0 @@
-﻿namespace Files.Enums
-{
-    public enum CloudProviders
-    {
-        OneDrive,
-        OneDriveCommercial,
-        Mega,
-        GoogleDrive,
-        DropBox,
-        AppleCloud,
-        AmazonDrive,
-        Box
-    }
-}
