+++ Files/Files.csproj
@@ -207,10 +207,8 @@
     <Compile Include="EventArguments\LayoutModeEventArgs.cs" />
     <Compile Include="Filesystem\CloudDrivesManager.cs" />
     <Compile Include="Helpers\AdaptiveLayoutHelpers.cs" />
-    <Compile Include="Filesystem\Cloud\Providers\OneDriveSharePointCloudProvider.cs" />
     <Compile Include="Filesystem\NetworkDrivesManager.cs" />
     <Compile Include="Helpers\AppServiceConnectionHelper.cs" />
-    <Compile Include="Filesystem\Cloud\Providers\AmazonDriveProvider.cs" />
     <Compile Include="Helpers\AppUpdater.cs" />
     <Compile Include="DataModels\DefaultLanguageModel.cs" />
     <Compile Include="Enums\FileOperationType.cs" />
@@ -412,17 +410,14 @@
     <Compile Include="Extensions\EnumerableExtensions.cs" />
     <Compile Include="Extensions\LinqExtensions.cs" />
     <Compile Include="Filesystem\Cloud\CloudDriveSyncStatusUI.cs" />
-    <Compile Include="Filesystem\Cloud\CloudProvider.cs" />
     <Compile Include="Filesystem\Cloud\Providers\AppleCloudProvider.cs" />
     <Compile Include="Filesystem\Cloud\Providers\BoxCloudProvider.cs" />
     <Compile Include="Enums\CloudDriveSyncStatus.cs" />
     <Compile Include="Filesystem\Cloud\CloudProviderController.cs" />
-    <Compile Include="Enums\CloudProviders.cs" />
     <Compile Include="Filesystem\Cloud\Providers\DropBoxCloudProvider.cs" />
     <Compile Include="Filesystem\Cloud\Providers\GoogleDriveCloudProvider.cs" />
     <Compile Include="Filesystem\Cloud\ICloudProviderDetector.cs" />
-    <Compile Include="Filesystem\Cloud\Providers\MegaCloudProvider.cs" />
-    <Compile Include="Filesystem\Cloud\Providers\OneDriveCloudProvider.cs" />
+    <Compile Include="Filesystem\Cloud\Providers\GenericCloudProvider.cs" />
     <Compile Include="Filesystem\FilesystemOperations\FilesystemItemType.cs" />
     <Compile Include="Filesystem\FilesystemOperations\FilesystemOperations.cs" />
     <Compile Include="Filesystem\FilesystemOperations\Helpers\FilesystemHelpers.cs" />
