+++ Files/Filesystem/Cloud/Providers/AppleCloudProvider.cs
@@ -1,4 +1,5 @@
-﻿using Files.Enums;
+﻿using Files.Common;
+using Files.Enums;
 using System;
 using System.Collections.Generic;
 using System.IO;
