+++ Files/Filesystem/Cloud/Providers/DropBoxCloudProvider.cs
@@ -1,4 +1,5 @@
-﻿using Files.Enums;
+﻿using Files.Common;
+using Files.Enums;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
